게시글 삭제 수정 게시글 작성자만 가능하게

+++ src/main/board-react-app/src/board/BoardView.jsx
@@ -1,7 +1,8 @@
-import React, {useEffect, useState} from "react";
+import React, {useContext, useEffect, useState} from "react";
 import {useNavigate, useParams} from "react-router-dom";
 import axios from "axios";
 import styled from "styled-components";
+import {AuthContext} from "../login/AuthContext";
 
 const BoardContainer = styled.div`
   padding: 20px;
@@ -72,16 +73,31 @@ const BackButton = styled.button`
 // 클릭한 게시글 id를 확인하기 위함 입니다.
 function BoardView() {
     const [boardView, setBoardView] = useState("");
+    const [currentUser, setCurrentUser] = useState("");
     const {id} = useParams(); // URL 파라미터에서 id를 추출합니다.
+
+    const {auth} = useContext(AuthContext); // 현재 로그인 한 사용자의 auth(인증 상태)를 가져옵니다.
 
     const navigate = useNavigate();
 
     // 컴포넌트가 마운트될 때 게시글 정보를 불러옵니다.
     useEffect(() => {
+
+        // 현재 로그인 된 사용자의 auth 값 가져오기
+        const loggedInUser = localStorage.getItem('auth');
+        if(loggedInUser) {
+            // 문자열로 저장된 인증 정보를 JavaScript 객체로 변환합니다.
+            setCurrentUser(JSON.parse(loggedInUser));
+
+            console.log('현재 로그인된 유저의 auth 정보입니다 : ', auth);
+        }
+
         const fetchBoardView = async () => {
             try {
                 const response = await axios.get(`/api/board/view/${id}`);
                 setBoardView(response.data); // 서버로부터 받은 데이터로 상태를 업데이트합니다.
+
+                console.log('현재 불러온 게시글 : ', response);
             } catch (error) {
                 console.error('게시글을 불러오는 데 실패했습니다.', error);
             }
@@ -93,6 +109,18 @@ function BoardView() {
 
     // 게시글 삭제 함수
     const handleDelete = async () => {
+
+        // 현재 로그인한 게시글 정보, 사용자 정보를 가져옵니다.
+        if (!boardView || !currentUser) {
+            alert('게시글 또는 사용자 정보를 불러올 수 없습니다.');
+            return;
+        }
+        // 관리자 또는 글 작성자의 경우만 삭제를 가능하게 합니다.
+        if (currentUser.username !== boardView.account_username && currentUser.role !== 'ROLE_ADMIN'){
+            alert('게시글 삭제 권한이 없습니다.');
+            return;
+        }
+
         if(window.confirm('정말로 글을 삭제하시겠습니까?')) {
             try {
                 await axios.delete(`/api/board/delete/${id}`);
@@ -106,14 +134,31 @@ function BoardView() {
         }
     }
 
+    const handleEdit = async () => {
+        // 현재 로그인한 게시글 정보, 사용자 정보를 가져옵니다.
+        if (!boardView || !currentUser) {
+            alert('게시글 또는 사용자 정보를 불러올 수 없습니다.');
+            return;
+        }
+        // 관리자 또는 글 작성자의 경우만 수정을 가능하게 합니다.
+        if (currentUser.username !== boardView.account_username && currentUser.role !== 'ROLE_ADMIN'){
+            alert('게시글 수정 권한이 없습니다.');
+            return;
+        }
+
+        navigate(`/board/editPage/${id}`);
+    }
+
     return (
         <div>
             <BoardContainer>
                 <Title>{boardView.title}</Title>
                 <Content>{boardView.content}</Content>
+
                 <BackButton onClick={() => navigate('/board/list')}>글 목록</BackButton>
+
                 <DeleteButton onClick={handleDelete}>글삭제</DeleteButton>
-                <EditButton onClick={() => navigate(`/board/editPage/${id}`)}>글 수정</EditButton>
+                <EditButton onClick={handleEdit}>글 수정</EditButton>
             </BoardContainer>
         </div>
     );
